requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==3.0.3
constantly==15.1.0
cryptography==41.0.1
cssselect2==0.7.0
daphne==3.0.2
Django==3.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
easy-thumbnails==2.8.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
lxml==4.9.2
Pillow==9.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pytz==2023.3
reportlab==4.0.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.3
webencodings==0.5.1
zope.interface==6.0